FOR_UPSTREAM [VPG]: drm/i915: Fix batch copying S3 bug
This patch fixes a bug in the commit below. It was checked in but w/ the
feature disabled due to crash issues seen on suspend and resume. The root
cause is that commit 1b01cf stores the batch pool lists inside
struct intel_ring_buffer. Unfortunately, these objects get destroyed on
suspend and recreated on resume, and so we lose the connection to the lists
which are supposed to persit across suspend and resume ops. The solution was
to store the lists in struct drm_i915_private instead.

	commit 1b01cf3facf0e46481e66ee884e329a4f18157ba
	Author: Brad Volkin <bradley.d.volkin@intel.com>
	Date:   Tue Oct 15 16:05:36 2013 -0700

    FOR_UPSTREAM [VPG]: drm/i915: Copy user batches to kernel buffer and submit

Change-Id: I89278310a55c0c6cff034b431826298404dfc687
For: AXIA-4496
OTC-Tracker: AXIA-4618
Signed-off-by: mhnguyen <michael.h.nguyen@intel.com>

Conflicts:

	drivers/gpu/drm/i915/i915_gem_execbuffer.c