REVERTME [VPG]: drm: removing DRM Master check on the DRM ioctls
Some ioctls (for example for mode setting) shall be allowed only for
the DRM master (whosever opens the DRM device file first will become
the DRM master). On Android side, SurfaceFlinger process is expected
to act like a DRM Master but, currently, DRM master check is failing
on the ioctls made by SurfaceFlinger process. This is because some
other service/process is opening the DRM device file & becoming a
master. For now, master check is not done.

Also for Android, currently all the Users are by default being marked
as authenticated. There is no X-Server like thing over here, which is
there in Linux desktop systems, which can act like a DRM master &
authenticate the clients. So currently all the DRM users/clients are
by default being marked as authenticated.

Originally, this patch was:
commit b72089dba5211f25a9e2450b3854c06bfa17a5ec
Author: akashgoe <akash.goel@intel.com>
Date:   Wed Sep 18 09:03:08 2013 +0530

    drm/i915: Adding Root permission check on the DRM/i915 ioctls

Issue: APDEV-983
Change-Id: I5cbfbbd6b7f60c1b77211f8666854f294ac72907
Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>
Signed-off-by: A.Sunil Kamath <sunil.kamath@intel.com>