lpc_ich: power button polling
Register with the ACPI button driver to provide polling of the power
button, so that correctly timed release can be detected and provided
to userspace.  All ICH southbridges support a PWRBTN_LVL bit in the
GEN_PMCON register, and more recent Bay Trail SoC's have an identical
bit with a slightly different access mechanism.

Change-Id: I0dee37d8eab03b4ad7c5a94502097bbee104943c
For: APDEV-575
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>