FOR_UPSTREAM [VPG] drm/edid: Fill PAR in AVI infoframe based on CEA mode list
Populate PAR in infoframe structure. If there is a user setting for PAR, then
that value is set. Else, value is taken from CEA mode list if VIC is found.
Else, PAR is calculated from resolution. If none of these conditions are
satisfied, PAR is NONE as per initialization.

Issue: APDEV-1251
OTC-Tracker: APDEV-1252
Change-Id: I78d5b6c16a31bfcd9f7b9d6634cef5c5f0740fe0
Signed-off-by: Vandana Kannan <vandana.kannan@intel.com>