FROM_UPSTREAM [VPG] drm/i915: Handle inactivating objects for all VMAs
This came from a patch called, "drm/i915: Move active to vma"

When moving an object to the inactive list, we do it for all VMs for
which the object is bound.

The primary difference from that patch is this time around we don't not
track 'active' per vma, but rather by object. Therefore, we only need
one unref.

Change-Id: I44eede1f3d938c8ecc8ff684a7b8acfcd92fe8aa
For: VIZ-3438
Upstream-repo: git://anongit.freedesktop.org/drm-intel
Upstream-commit: feb822cfc2540c2d2df7827f40991aa2f86f1130
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-on: https://android.intel.com/163403
Reviewed-by: Landowski, Piotr <piotr.landowski@intel.com>
Reviewed-by: Odeberg, Hans <hans.odeberg@intel.com>
Tested-by: Odeberg, Hans <hans.odeberg@intel.com>