FROM_UPSTREAM [VPG] drm/i915: Do not call retire_requests from wait_for_rendering
A common issue we have is that retiring requests causes recursion
through GTT manipulation or page table manipulation which we can only
handle at very specific points. However, to maintain internal
consistency (enforced through our sanity checks on write_domain at
various points in the GEM object lifecycle) we do need to retire the
object prior to marking it with a new write_domain, and also clear the
write_domain for the implicit flush following a batch.

Note that this then allows the unbound objects to still be on the active
lists, and so care must be taken when removing objects from unbound lists
(similar to the caveats we face processing the bound lists).

Change-Id: I76d299b8cb42851e333d06df7dd1a28252ef0781
Upstream-repo: http://lists.freedesktop.org/archives/intel-gfx/2014-January/038989.html
For: VIZ-3438
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-on: https://android.intel.com/163405
Reviewed-by: Landowski, Piotr <piotr.landowski@intel.com>
Reviewed-by: Odeberg, Hans <hans.odeberg@intel.com>
Tested-by: Odeberg, Hans <hans.odeberg@intel.com>