[REVERTME] Dell xps quirk
The camera on Dell xps sometimes sends payloads larger than the
maximum reported payloadsize (dwMaxPayloadTransferSize). For this
reason the driver assumes it has found the end of the payload and
expects the next urb to be the start of a new frame, which it isn't.
This is a workaround for this issue.

Issue: APDEV-865
Signed-off-by: Ryan, Mark D <mark.d.ryan@intel.com>
Signed-off-by: Adriana Reus <adriana.reus@intel.com>

Change-Id: I2433f8821f659c1b7af34a082f05a4bafdff1386